Dependency tests for ReactiveTutor. Minor refactoring to ensure that Singleton instances can be used by multiple tests

+++ TutoringModule/ReactiveTutor/Its.TutoringModule.ReactiveTutorTest/ReactiveTutorTest.cs
@@ -10,9 +10,13 @@ using NUnit.Framework;
 namespace Its.TutoringModule.ReactiveTutorTest
 {
     [TestFixture]
+    [Parallelizable(ParallelScope.Self)]
     public class ReactiveTutorTest
     {
-        public ITutorConfig InitializeTest(string testName)
+        private static ITutorConfig config;
+        
+        [OneTimeSetUp]
+        public void InitFixture()
         {
             string runtimePath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
             string assetPath = runtimePath + "/TestAssets/";
@@ -24,40 +28,53 @@ namespace Its.TutoringModule.ReactiveTutorTest
             Dictionary<string, string> settings = new Dictionary<string, string>()
             {
                 {"OntologyPath", assetPath + "Ontology/"},
-                {"LogsPath", assetPath + testName + "/Logs/"},
-                {"DomainConfigurationPath", assetPath + testName + "/DomainConf/"},
-                {"WorldConfigurationPath", assetPath + testName + "/WorldConf/"}
+                {"LogsPath", assetPath + "/Logs/"},
+                {"DomainConfigurationPath", assetPath + "/DomainConf/"},
+                {"WorldConfigurationPath", assetPath + "/WorldConf/"}
             };
             
-            ITutorConfig config = new DefaultTutorConfig(template, settings);
+            config = new DefaultTutorConfig(template, settings);
             
             // Wipe any previous logs if present
-            Directory.Delete(config.LogsPath, true);
-            Directory.Delete(config.DomainConfigurationPath + "/Logs", true);
+            if (Directory.Exists(config.LogsPath))
+            {
+                Directory.Delete(config.LogsPath, true);
+            }
 
-            return config;
+            if (Directory.Exists(config.DomainConfigurationPath + "/Logs"))
+            {
+                Directory.Delete(config.DomainConfigurationPath + "/Logs", true);
+            }
+        }
+        
+        public void InitTest(string domainName)
+        { 
+            // Wipe any previous logs if present
+            if (Directory.Exists(config.LogsPath + domainName))
+            {
+                Directory.Delete(config.LogsPath + domainName, true);
+            }
+
+            if (Directory.Exists(config.DomainConfigurationPath + "/Logs/" + domainName))
+            {
+                Directory.Delete(config.DomainConfigurationPath + "/Logs/" + domainName, true);
+            }
         }
         
         [Test]
         public void TestHappyPath()
         {
             // Arrange
-            string testName = "HappyPath";
-            ITutorConfig config = InitializeTest(testName);
+            string domainName = "HappyPath";
+            InitTest(domainName);
             
             TutorFactory tf = TutorFactory.Instance();
-            ITutor tutor = tf.CreateReactiveTutor(testName, config);
+            ITutor tutor = tf.CreateReactiveTutor(domainName, config);
 
             string studentKey = "1";
-//            try
-//            {
-//                tutor.CreateStudent(studentKey, "Test", "Test", "Test", "Test");
-//            }
-//            catch (Exception e)
-//            {
-//                // Most likely stuent with such ID already exists.
-//            }
-//            tutor.CreateStudent(studentKey, "Test", "Test", "Test", "Test");
+            
+            // Clean any previous test run logs from memory
+            tutor.ResetPractice(domainName, studentKey);
             
             // Act
             // Assert
@@ -67,30 +84,193 @@ namespace Its.TutoringModule.ReactiveTutorTest
             Dictionary<string, List<string>> messages;
 
             // action1
-            code = tutor.Validate("action1", testName, studentKey, "obj", out errors);
-            Assert.AreEqual(code, 1);
-            Assert.AreEqual(errors.Count, 0);
-            messages = tutor.GetTutorMessages("action1", testName, studentKey);
+            code = tutor.Validate("action1", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action1", domainName, studentKey);
             Assert.AreEqual(2, messages.Keys.Count);
             Assert.AreEqual("action1", messages["confirmation"][0]);
             Assert.AreEqual("action2", messages["tutor"][0]);
             
             // action2
-            code = tutor.Validate("action2", testName, studentKey, "obj", out errors);
-            Assert.AreEqual(code, 1);
-            Assert.AreEqual(errors.Count, 0);
-            messages = tutor.GetTutorMessages("action2", testName, studentKey);
+            // Bug present in displaying confirmation message - control flag in spreadsheet is currently ignored
+            code = tutor.Validate("action2", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action2", domainName, studentKey);
             Assert.AreEqual(2, messages.Keys.Count);
             Assert.AreEqual("action2", messages["confirmation"][0]);
             Assert.AreEqual("action3", messages["tutor"][0]);
             
             // action3
-            code = tutor.Validate("action3", testName, studentKey, "obj", out errors);
-            Assert.AreEqual(code, 1);
-            Assert.AreEqual(errors.Count, 0);
-            messages = tutor.GetTutorMessages("action3", testName, studentKey);
+            code = tutor.Validate("action3", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action3", domainName, studentKey);
             Assert.AreEqual(1, messages.Keys.Count);
             Assert.AreEqual("action3", messages["confirmation"][0]);
+        }
+        
+        [Test]
+        public void TestSimpleDependency()
+        {
+            // Arrange
+            string domainName = "SimpleDependency";
+            InitTest(domainName);
+            
+            TutorFactory tf = TutorFactory.Instance();
+            ITutor tutor = tf.CreateReactiveTutor(domainName, config);
+
+            string studentKey = "1";
+            
+            // Clean any previous test run logs from memory
+            tutor.ResetPractice(domainName, studentKey);
+            
+            // Act
+            // Assert
+
+            int code;
+            List<string> errors;
+            Dictionary<string, List<string>> messages;
+
+            // action1
+            code = tutor.Validate("action1", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action1", domainName, studentKey);
+            Assert.AreEqual(2, messages.Keys.Count);
+            Assert.AreEqual("action1", messages["confirmation"][0]);
+            Assert.AreEqual("action2", messages["tutor"][0]);
+            
+            // action2
+            code = tutor.Validate("action2", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action2", domainName, studentKey);
+            Assert.AreEqual(1, messages.Keys.Count);
+            Assert.AreEqual("action2", messages["confirmation"][0]);
+        }
+        
+        [Test]
+        public void TestSimpleDependencyError()
+        {
+            // Arrange
+            string domainName = "SimpleDependency";
+            InitTest(domainName);
+            
+            TutorFactory tf = TutorFactory.Instance();
+            ITutor tutor = tf.CreateReactiveTutor(domainName, config);
+
+            string studentKey = "1";
+            
+            // Clean any previous test run logs from memory
+            tutor.ResetPractice(domainName, studentKey);
+            
+            // Act
+            // Assert
+
+            int code;
+            List<string> errors;
+            Dictionary<string, List<string>> messages;
+
+            // action1
+            // Skip to cause validation error when performing action2
+            
+            // action2
+            code = tutor.Validate("action2", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(0, code);
+            Assert.AreEqual(1, errors.Count);
+            Assert.AreEqual("action1", errors[0]);
+        }
+        
+        [Test]
+        public void TestOptDependency()
+        {
+            // Arrange
+            string domainName = "OptDependency";
+            InitTest(domainName);
+            
+            TutorFactory tf = TutorFactory.Instance();
+            ITutor tutor = tf.CreateReactiveTutor(domainName, config);
+
+            string studentKey = "1";
+            
+            // Clean any previous test run logs from memory
+            tutor.ResetPractice(domainName, studentKey);
+            
+            // Act
+            // Assert
+
+            int code;
+            List<string> errors;
+            Dictionary<string, List<string>> messages;
+
+            // action1
+            code = tutor.Validate("action1", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action1", domainName, studentKey);
+            Assert.AreEqual(2, messages.Keys.Count);
+            Assert.AreEqual("action1", messages["confirmation"][0]);
+            Assert.AreEqual("action2", messages["tutor"][0]);
+            
+            // action2
+            code = tutor.Validate("action2", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action2", domainName, studentKey);
+            Assert.AreEqual(2, messages.Keys.Count);
+            Assert.AreEqual("action2", messages["confirmation"][0]);
+            Assert.AreEqual("action3", messages["tutor"][0]);
+            
+            // action3
+            code = tutor.Validate("action3", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action3", domainName, studentKey);
+            Assert.AreEqual(1, messages.Keys.Count);
+            Assert.AreEqual("action3", messages["confirmation"][0]);
+        }
+        
+        [Test]
+        public void TestOptDependencyError()
+        {
+            // Arrange
+            string domainName = "OptDependency";
+            InitTest(domainName);
+            
+            TutorFactory tf = TutorFactory.Instance();
+            ITutor tutor = tf.CreateReactiveTutor(domainName, config);
+
+            string studentKey = "1";
+            
+            // Clean any previous test run logs from memory
+            tutor.ResetPractice(domainName, studentKey);
+            
+            // Act
+            // Assert
+
+            int code;
+            List<string> errors;
+            Dictionary<string, List<string>> messages;
+
+            // action1
+            code = tutor.Validate("action1", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(1, code);
+            Assert.AreEqual(0, errors.Count);
+            messages = tutor.GetTutorMessages("action1", domainName, studentKey);
+            Assert.AreEqual(2, messages.Keys.Count);
+            Assert.AreEqual("action1", messages["confirmation"][0]);
+            Assert.AreEqual("action2", messages["tutor"][0]);
+            
+            // action2
+            // skipping
+            
+            // action3
+            code = tutor.Validate("action3", domainName, studentKey, "obj", out errors);
+            Assert.AreEqual(0, code);
+            Assert.AreEqual(1, errors.Count);
+            Assert.AreEqual("action2", errors[0]);
         }
     }
 }
